# String#split seems to be faster than capturing digits with Regexp