@amatsuda 2023-01-06
Ruby 2.3 tests are failing, because `Regexp#match?` was not implemented in that version yet. Should we still support that version here?