@hsbt 2023-01-10
I'm +1 to drop to support Ruby 2.3 for performance improvement.